requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
scipy>=1.11.0
pygame>=2.5.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
scikit-learn>=1.3.0
filterpy>=1.4.5


